+++ rust/mlt-core/src/utils/extensions.rs
@@ -17,31 +17,6 @@ impl<T> SetOptionOnce<T> for Option<T> {
     }
 }
 
-pub trait AsUsize: Eq + Copy {
-    fn as_usize(&self) -> usize;
-}
-
-impl AsUsize for usize {
-    #[inline]
-    fn as_usize(&self) -> usize {
-        *self
-    }
-}
-
-impl AsUsize for u32 {
-    #[inline]
-    fn as_usize(&self) -> usize {
-        const _: () = {
-            // Some day Rust may support usize smaller than u32?
-            assert!(
-                size_of::<u32>() <= size_of::<usize>(),
-                "usize must be able to hold all u32 values"
-            );
-        };
-        usize::try_from(*self).unwrap()
-    }
-}
-
 /// Perform checked addition of two values, returning an error if any overflow occurs.
 #[inline]
 pub fn checked_sum2<T: CheckedAdd + Copy>(v1: T, v2: T) -> MltResult<T> {