+++ rust/mlt-synthetics/src/main.rs
@@ -325,6 +325,16 @@ fn generate_geometry(w: &mut SynthWriter) {
         .geo(MultiLineString(vec![line1(), line2()]))
         .write(w, "multiline");
 
+    // See https://github.com/maplibre/maplibre-gl-js/issues/7659
+    //
+    // There was a mismatch between what rust encoded and what ts decoded
+    // in the case of single element geometry streams.
+    geo_varint()
+        .meta(E::delta_varint())
+        .no_rings(E::rle_varint())
+        .geo(MultiLineString(vec![line1(), line2()]))
+        .write(w, "multiline_meta_delta-rust");
+
     // Split the Morton curve into two halves to form a MultiLineString with Morton encoding.
     let mline1 = LineString::new(mc[..half].to_vec());
     let mline2 = LineString::new(mc[half..].to_vec());