+++ ts/src/decoding/integerStreamDecoder.ts
@@ -122,6 +122,9 @@ export function decodeUnsignedConstInt32Stream(
     const values = decodePhysicalLevelTechnique(data, offset, streamMetadata);
 
     if (values.length === 1) {
+        if (streamMetadata.logicalLevelTechnique1 === LogicalLevelTechnique.DELTA) {
+            return decodeZigZagInt32Value(values[0]);
+        }
         return values[0];
     }