+++ rust/mlt-core/CHANGELOG.md
@@ -7,6 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0](https://github.com/maplibre/maplibre-tile-spec/compare/rust-mlt-core-v0.9.3...rust-mlt-core-v0.10.0) - 2026-06-08
+
+### Added
+
+- *(py)* simple, geojson based encoding ([#1416](https://github.com/maplibre/maplibre-tile-spec/pull/1416))
+
+### Other
+
+- *(rust)* use `usize_cast` for casting ([#1414](https://github.com/maplibre/maplibre-tile-spec/pull/1414))
+- *(rust)* migrate to fast-mvt crate ([#1415](https://github.com/maplibre/maplibre-tile-spec/pull/1415))
+
+## [0.9.3](https://github.com/maplibre/maplibre-tile-spec/compare/rust-mlt-core-v0.9.2...rust-mlt-core-v0.9.3) - 2026-05-15
+
+### Other
+
+- update Cargo.toml dependencies
+
+## [0.9.2](https://github.com/maplibre/maplibre-tile-spec/compare/rust-mlt-core-v0.9.1...rust-mlt-core-v0.9.2) - 2026-05-14
+
+### Other
+
+- make errors more accurate ([#1370](https://github.com/maplibre/maplibre-tile-spec/pull/1370))
+
+## [0.9.1](https://github.com/maplibre/maplibre-tile-spec/compare/rust-mlt-core-v0.9.0...rust-mlt-core-v0.9.1) - 2026-05-05
+
+### Added
+
+- *(rust)* MLT -> MVT write support ([#1369](https://github.com/maplibre/maplibre-tile-spec/pull/1369))
+
 ## [0.9.0](https://github.com/maplibre/maplibre-tile-spec/compare/rust-mlt-core-v0.8.0...rust-mlt-core-v0.9.0) - 2026-04-29
 
 ### Added